requirements.txt
streamlit
pdf2image
pytesseract
spacy
pillow
scikit-learn
transformers
torch